refactor: move CertificatesPreview component to its own file
- Move CertificatesPreview to certificates-preview.tsx
- Update the import paths to this component
